Refactor fetch post
