+++ src/post/type/like.interface.ts
@@ -1,0 +1,5 @@
+export interface ILikes {
+  id: number;
+  postId: number;
+  userId: number;
+}
