+++ src/store/modules/comments/index.ts
@@ -1,9 +1,11 @@
 import Axios from "axios";
-import { IComment } from '@/types/comments';
+import { IComment } from "@/post/type/comment.interface";
 
 const actions = {
   fetchComments: (context: any, postsId: number) => {
-    return Axios.get(`http://localhost:3000/comments/${postsId}`).then(res => res.data);
+    return Axios.get(`http://localhost:3000/comments/${postsId}`).then(
+      res => res.data
+    );
   },
   writeComment: (context: any, comment: IComment) => {
     return Axios.post(`http://localhost:3000/comments`, comment);
@@ -14,5 +16,5 @@ const actions = {
 };
 
 export default {
-  actions,
+  actions
 };
