--- src/types/likes.ts
@@ -1,8 +1,0 @@
-import { IPost } from "../post/type/post.interface";
-import { IUser } from "./user";
-
-export interface ILikes {
-  id: number;
-  postId: number;
-  userId: number;
-}
