+++ src/types/user.ts
@@ -1,6 +1,7 @@
 export interface IUser {
   id: number;
   email: string;
+  name: string;
   password: string;
   createdAt: Date;
 }
